add Data_File and update Message Uploader

+++ services/uploader.js
@@ -2,6 +2,7 @@ var Formidable = require('formidable');
 var Path = require('path');
 var fs = require('fs');
 var UFile = require('../models/ufile');
+var DataFile = require('../models/data_file.js');
 var config    = require('../config/config');
 var IM = require('imagemagick');
 
@@ -24,7 +25,7 @@ module.exports.upload = function (req) {
             var size = files.size;
             var path = files.path;
 
-            if(!fields.id)
+            if(!fields.id && !fields.device_id)
             {
                 req.reply(false);
                 fs.unlink(path);
@@ -55,8 +56,13 @@ module.exports.upload = function (req) {
             {
                 req.reply(relativePath);
             }
-            var uf = new UFile({uploader:fields.id, type:type, size:size, path:path});
-            uf.save();
+            if(fields.id){
+                var uf = new UFile({uploader:fields.id, type:type, size:size, path:path});
+                uf.save();
+            }else if(fields.device_id){
+                var df = new DataFile({device_id: fields.device_id, size: size, path: path});
+                df.save();
+            }
         }
     });
